pyproject.toml
[project]
name = "non-threaded"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "create-agent>=0.0.3",
    "fastapi>=0.128.0",
    "fsspec>=2025.12.0",
    "langchain>=1.2.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "pandas>=2.3.3",
    "pip>=25.3",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
